Refactor debugger between client/server
